# Conflicts:
#	WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/include/vector_tiles/SLDSymbolizers.h
#	WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/src/vector_tiles/SLDStyleSet.m


Former-commit-id: bdac084ead57d01f3a90d37138a9c3ada1f1a163

+++ WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/src/vector_tiles/vector_tile.pb.cpp
@@ -390,7 +390,7 @@ void Tile_Value::SerializeWithCachedSizes(
       8, 536870912, output);
 
   output->WriteRaw(unknown_fields().data(),
-                   unknown_fields().size());
+                   (int) unknown_fields().size());
   // @@protoc_insertion_point(serialize_end:vector_tile.Tile.Value)
 }
 
@@ -772,7 +772,7 @@ void Tile_Feature::SerializeWithCachedSizes(
   }
 
   output->WriteRaw(unknown_fields().data(),
-                   unknown_fields().size());
+                   (int) unknown_fields().size());
   // @@protoc_insertion_point(serialize_end:vector_tile.Tile.Feature)
 }
 
@@ -1162,7 +1162,7 @@ void Tile_Layer::SerializeWithCachedSizes(
       16, 536870912, output);
 
   output->WriteRaw(unknown_fields().data(),
-                   unknown_fields().size());
+                   (int) unknown_fields().size());
   // @@protoc_insertion_point(serialize_end:vector_tile.Tile.Layer)
 }
 
@@ -1451,7 +1451,7 @@ void Tile::SerializeWithCachedSizes(
       16, 8192, output);
 
   output->WriteRaw(unknown_fields().data(),
-                   unknown_fields().size());
+                   (int) unknown_fields().size());
   // @@protoc_insertion_point(serialize_end:vector_tile.Tile)
 }
